% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminLogs.R
\name{adminLogsUI}
\alias{adminLogsUI}
\alias{adminLogs}
\title{Admin log file viewer shiny module}
\usage{
adminLogsUI(id)

adminLogs(input, output, session, config, path = "/var/log/shiny-server")
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{config}{\pkg{SpaDES.shiny} app configuration object (e.g., from \code{\link{readConfig}})}

\item{path}{file path of the server log files (usually \file{/var/log/shiny-server})}
}
\description{
View app-specific logs at a specific path (default is \file{/var/log/shiny-server}).
}
\details{
Generates a \code{fluidRow} consisting of two columns: one containing a file selector
dropdown menu and the other displaying the text output from the selected log file.
}
\author{
Alex Chubaty
}
