% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapList.R
\name{appendMapList}
\alias{appendMapList}
\title{Append a spatial object to mapList}
\usage{
appendMapList(object, ...)
}
\description{
Append a spatial object to mapList
}
\examples{
library(sp)
library(raster)
coords <- structure(c(-122.98, -116.1, -99.2, -106, -122.98,
                      59.9, 65.73, 63.58, 54.79, 59.9),
                    .Dim = c(5L, 2L))
Sr1 <- Polygon(coords)
Srs1 <- Polygons(list(Sr1), "s1")
StudyArea <- SpatialPolygons(list(Srs1), 1L)
crs(StudyArea) <- "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

ml <- new("mapList")
appendMapList(StudyArea, ml, studyArea = TRUE, layerName = "newPoly")

}
