% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authGoogle.R
\name{authGoogleUI}
\alias{authGoogleUI}
\alias{authGoogle}
\title{Authorize a user using Google Drive account}
\usage{
authGoogleUI(id)

authGoogle(input, output, session, appURL, authUsers, icon = "google")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{appURL}{URL to the hosted app.}

\item{authUsers}{A character vector of authorized user email addresses.}

\item{icon}{Default \code{"google"}. Name of icon to display beside the login button.
Use of an icon with \pkg{shinydashboard} may produce undesired results.
Disable use of the icon with \code{icon = NULL}.}
}
\value{
A list of the user's Google profile details (name, email, etc.)
}
\description{
Logs a user into their Google account for authorized use of the app.
Allows app content to be made conditionally available to 3 classes of user:
}
\details{
1. Anonymous (non-logged in) users;
2. Logged in users who are not autheticated against a whitelist;
3. Logged in and authenticated users.

This allows authentication to be managed via Google rather than locally on the shiny server.
Only the whitelist of approved users is kept locally.

Login status is stored in \code{session$userData$userLoggedIn()}, a reactive value.
Authentication status is stored in \code{session$userData$userAuthorized()}, a reactive value.
}
\note{
Based on \code{googleAuthR::googleAuth} and \code{googleAuthR::googleAuthUI}.
}
\section{Additional requirements}{

Your \file{global.R} file should set the following options:
\enumerate{
  \item \code{googleAuthR.scopes.selected}:
  \preformatted{c("https://www.googleapis.com/auth/userinfo.email",
  "https://www.googleapis.com/auth/userinfo.profile")}
  \item \code{googleAuthR.webapp.client_id}: your Google app oauth id
  \item \code{googleAuthR.webapp.client_secret}: your Google app oauth "secret"
}

Be sure to also set \code{appURL} and \code{authUsers} in \file{global.R}.

See the authentication vignette (\code{vignette("authentication", "SpaDES.shiny")}).
}

\author{
Alex Chubaty
}
