% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cloneApp}
\alias{cloneApp}
\title{Clone an existing \pkg{SpaDES.shiny} app to a new directory}
\usage{
cloneApp(from, to, symlinks = FALSE)
}
\arguments{
\item{from}{Directory of the app from which to copy/clone.}

\item{to}{New app directory to be created.}

\item{symlinks}{Logical indicating whether symlinks should be used for large
directories, including \file{app_data/}, \file{modules/},
and \file{packrat/}. Using symlinks speeds up app cloning.}
}
\value{
None. Invoked for the side-effect of copyin an existing app.
}
\description{
Setup the necessary directory structure for a new app, if not already present.
Useful for copying an existing app for development or deployment purposes.
}
\details{
This does not copy the app \file{cache/} nor \file{output/} directories.
}
\author{
Alex Chubaty
}
