% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpMod2.R
\name{clumpMod2UI}
\alias{clumpMod2UI}
\alias{clumpMod2}
\title{Clump module}
\usage{
clumpMod2UI(id)

clumpMod2(input, output, session, tsf, vtm, currentPolygon, cl, ageClasses,
  patchSize, sizeInHa, cacheRepo, largePatchesFn, countNumPatches)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's server function}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{tsf}{...}

\item{vtm}{...}

\item{currentPolygon}{...}

\item{cl}{...}

\item{ageClasses}{...}

\item{patchSize}{...}

\item{sizeInHa}{...}

\item{cacheRepo}{...}

\item{largePatchesFn}{...}

\item{countNumPatches}{...}
}
\value{
Shiny module UI.

Shiny module UI.
}
\description{
Shiny module used ...

Shiny module used to ... # TODO: documentation needed
}
\author{
Damian Rodziewicz

Damian Rodziewicz
}
