% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{writeConfig}
\alias{writeConfig}
\title{Write yaml config file}
\usage{
writeConfig(config, file)
}
\arguments{
\item{config}{List of configuration variables to write to file}

\item{file}{Path to yaml config file}
}
\description{
Write yaml config file
}
