% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{dataInfoUI}
\alias{dataInfoUI}
\alias{dataInfo}
\title{Module data sources information module}
\usage{
dataInfoUI(id)

dataInfo(input, output, session, sim)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
Display the rendered Rmarkdown content of the module's data description file
(\file{moduleName/moduleName_data.Rmd}).
}
\author{
Alex Chubaty
}
