% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{debuggingUI}
\alias{debuggingUI}
\alias{debugging}
\title{Generate Debugging Tab}
\usage{
debuggingUI(id)

debugging(input, output, session, config, simID)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{config}{inSpaDES app configuration (e.g., from \code{\link{readConfig}})}

\item{simID}{The simulation number}
}
\description{
View app-specific logs at a specific path (default is \file{/var/log/shiny-server}).
}
\details{
Generates a tab consisting of the text output from the log file.
}
\author{
Olivia Sung, Alex Chubaty, Alex Tso
}
