% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDataTable.R
\name{displayDataTableUI}
\alias{displayDataTableUI}
\alias{displayDataTable}
\title{Display Data Table Module}
\usage{
displayDataTableUI(id, ...)

displayDataTable(input, output, session, data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{...}{Additional parameters passed to \code{link[DT]{dataTableOutput}}}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{data}{Data table which should be displayed.}
}
\value{
None. Invoked for the side-effect of generating UI for plot

None. Invoked for the side-effect of rendering bar plot.
}
\description{
Shiny module which displays a data table.
}
