% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportToFile}
\alias{exportToFile}
\title{Export to file}
\usage{
exportToFile(file, data, device)
}
\arguments{
\item{file}{shiny server input object}

\item{data}{shiny server output object}

\item{device}{shiny server session object}
}
\description{
Export to file
}
