% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footers.R
\name{copyrightFooter}
\alias{copyrightFooter}
\alias{sidebarFooter}
\title{Generate HTML footers for body or sidebar}
\usage{
copyrightFooter(copyrightInfo)

sidebarFooter()
}
\arguments{
\item{copyrightInfo}{Character string containing your app's copyright info.}
}
\description{
App copyright info can be inserted into a footer in the dashboard body using
\code{copyrightFooter}.
}
\author{
Alex Chubaty

Alex Chubaty
}
