% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generateSpadesShinyUI}
\alias{generateSpadesShinyUI}
\title{Generate and save ui.R file.}
\usage{
generateSpadesShinyUI(appDir, appMetadata)
}
\arguments{
\item{appDir}{The directory path to use for the new app.}

\item{appMetadata}{Application metadata.}
}
\value{
None. Invoked for the side-effect of writing generated template to file.
}
\description{
Generate and save ui.R file.
}
\author{
Damian Rodziewicz
}
