% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{.generateUI}
\alias{.generateUI}
\title{Generate UI (internal)}
\usage{
.generateUI(uiType, categoriesValues, ns)
}
\arguments{
\item{uiType}{Character indicating the UI type; currently one of \code{"box"} or \code{"tab"}.}

\item{categoriesValue}{Single column \code{data.table} containing the names of the categories.}
}
\description{
Generate UI (internal)
}
\keyword{internal}
