% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{getModuleById}
\alias{getModuleById}
\title{Retrieve a module metadata from modules tibble.}
\usage{
getModuleById(modules, moduleId)
}
\arguments{
\item{modules}{Tibble with modules metadata. Tibble format: type, name, id, parameters.}

\item{moduleId}{Id of the module to retrieve.}
}
\value{
Tibble containing the module if it was found. Empty tibble otherwise.
}
\description{
Retrieve a module metadata from modules tibble.
}
\author{
Damian Rodziewicz
}
