% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{getModuleByName}
\alias{getModuleByName}
\title{Retrieve a module metadata from modules tibble.}
\usage{
getModuleByName(modules, moduleName)
}
\arguments{
\item{modules}{Tibble with modules metadata. Tibble format: type, name, uiParameters, serverParameters.}

\item{moduleName}{Name of the module to find.}
}
\value{
Tibble containing the module if it was found. Empty tibble otherwise.
}
\description{
Retrieve a module metadata from modules tibble.
}
\author{
Damian Rodziewicz
}
