% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authGoogle.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Get a uner's Google User info}
\usage{
getUserInfo(id = "me")
}
\description{
Get the Google user's name and email address.
}
\details{
Based on \code{\link[googleID]{get_user_info}}, but uses the Drive API instead of Google+
}
