% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramForRaster.R
\name{histogramForRaster}
\alias{histogramForRaster}
\title{Histogram for Raster Shiny Module}
\usage{
histogramForRaster(input, output, session, raster)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{raster}{Reactive value containing raster}
}
\description{

}
