% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramForRaster.R
\name{histogramForRasterUI}
\alias{histogramForRasterUI}
\title{Histogram for Raster Shiny Module}
\usage{
histogramForRasterUI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function}
}
\description{
This creates UI for shiny module creating histogram of data contained in raste
}
