% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramUI}
\alias{histogramUI}
\title{Histogram Module UI Function}
\usage{
histogramUI(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{...}{Additional parameters passed to \code{link[shiny]{plotOutput}}}
}
\value{
None. Invoked for the side-effect of generating UI for plot
}
\description{
UI function of a shiny module which creates a histogram
             using barplot based on the data received
}
