% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOutputs.R
\name{initialMapUI}
\alias{initialMapUI}
\alias{initialMap}
\title{Initial map module}
\usage{
initialMapUI(id)

initialMap(input, output, session, sim, mapID)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}

\item{mapID}{Character string indicating the name of the map in the \code{simList} object.}
}
\description{
TO DO: need description
}
\author{
Alex Chubaty
}
