% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largePatches.R
\name{largePatchesUI}
\alias{largePatchesUI}
\alias{largePatches}
\title{Large patches (shiny module)}
\usage{
largePatchesUI(id)

largePatches(session, input, output, numberOfSimulationTimes, clumpMod2Args)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{session}{Shiny server session object.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{numberOfSimulationTimes}{How many simulation time stamps there are.}

\item{clumpMod2Args}{List containing named arguments passed to \code{clumpMod2}.}
}
\value{
shiny module UI.

Shiny module server function.
}
\description{
Create summary for large patches function.
}
\author{
Mateusz Wyszynski

Mateusz Wyszynski
}
