% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libInit.R
\name{libInit}
\alias{libInit}
\title{Initialize a \pkg{SpaDES.shiny} app's \pkg{packrat} library}
\usage{
libInit(APP_DIR)
}
\arguments{
\item{APP_DIR}{File path giving the application directory}
}
\description{
Spawns a new \code{Rscript} process to setup an app's package library using
\pkg{packrat}. It initializes the packrat library inside the application
directory, and proceeds to install all \pkg{SpaDES.shiny} package dependencies,
plus \code{devtools}.
\strong{WARNING: this may take a while to complete!}
}
\note{
This spawns and runs in another R process, during which time your
current R process will wait, and RStudio may appear to hang while it's running.
}
\author{
Alex Chubaty
}
