% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapList.R
\docType{class}
\name{mapList-class}
\alias{mapList-class}
\alias{mapList}
\title{The \code{mapList} class}
\description{
Contains a common system for organzing vector and raster
layers, principly for use with leaflet and shiny.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{List of character names specifying which modules to load.}

\item{\code{maps}}{Named list of map-type objects (e.g., \code{sf}, \code{Raster*},
\code{Spatial*}.}

\item{\code{rasterTiles}}{Paths to rasterTiles}

\item{\code{CRS}}{The common crs of all layers}

\item{\code{rasterTemplate}}{The empty raster with all metadata filled}

\item{\code{analyses}}{A data.table or data.frame of the types of analyses to perform}

\item{\code{analysesData}}{A data.table or data.frame of the results of the analyses}
}}

