% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{moduleParamsUI}
\alias{moduleParamsUI}
\alias{moduleParams}
\title{\code{SpaDES} module parameter values module}
\usage{
moduleParamsUI(id)

moduleParams(input, output, session, sim)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
  UI function

TO DO: needs documentation}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
\code{SpaDES} module parameter values module
}
\author{
Alex Chubaty
}
