% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{newAppDeprecated}
\alias{newAppDeprecated}
\title{Create a new \pkg{SpaDES.shiny} app}
\usage{
newAppDeprecated(APP_DIR, ...)
}
\arguments{
\item{APP_DIR}{The directory path to use for the new app.}

\item{...}{Additional arguments to populate app's \file{_config.yml} file.}
}
\value{
Invisibly return the config list.
}
\description{
Setup the necessary directory structure for a new app, and create a new app
based on a template.
}
\details{
The optional argument \code{RLIB_DIR} may be set to override the default of using a
\pkg{packrat} library managing the app's package dependencies.
}
\author{
Alex Chubaty
}
