% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{newConfig}
\alias{newConfig}
\title{Create a new app config file}
\usage{
newConfig(APP_DIR, ...)
}
\arguments{
\item{APP_DIR}{The directory path to use for the new app.}

\item{...}{Additional arguments to populate app's \file{_config.yml} file.}
}
\description{
Create a new app config file
}
\author{
Alex Chubaty
}
