% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonList.R
\name{polygonList}
\alias{polygonList}
\title{Create a new \code{polygonList} object}
\usage{
polygonList(studyArea, ...)
}
\arguments{
\item{studyArea}{A template \code{Spatial*} object whose projection, extent,
etc. will be used for the polygons being added.}

\item{...}{\code{SpatialPolygonsDataFrame} objects to be added.}
}
\description{
Create a new \code{polygonList} object
}
