% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{pkgDeps}
\alias{pkgDeps}
\alias{pkgVers}
\alias{pkgAuthors}
\title{Get installed package dependencies}
\usage{
pkgDeps(x, type = NULL)

pkgVers(x, lib.loc = NULL)

pkgAuthors(x)
}
\arguments{
\item{x}{An installed package name.}

\item{type}{One of \code{"Depends"}, \code{"Imports"}, \code{"LinkingTo"}, \code{"Suggests"},
\code{"Enhances"}, or \code{"all"}. Default is \code{NULL}, which returns all.}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL.
The default value of NULL corresponds to all libraries currently known.
If the default is used, the loaded packages and namespaces are searched
before the libraries.}
}
\value{
A character vector of package dependencies.

A character vector of package dependencies.

A person vector of package authors.
}
\description{
This is simply a wrapper around \code{tools::package_dependencies}

This is simply a wrapper around \code{utils::packageVersion} to 1) allow a
vector of package names to be passed as an argument; 2) to warn instead of error
when a package is not installed; and 3) return versions as character strings.
}
\author{
Alex Chubaty
}
