% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{pkgInformationUI}
\alias{pkgInformationUI}
\alias{pkgInformation}
\title{Generate Package Dependency Version Information}
\usage{
pkgInformationUI(id)

pkgInformation(input, output, session, config)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{config}{\pkg{SpaDES.shiny} app configuration object (e.g., from \code{\link{readConfig}})}
}
\description{
Generate Package Dependency Version Information
}
\details{
Generates a box containing a data table with \pkg{SpaDES.shiny} package
dependency version information (i.e., all the packages required or used by \pkg{SpaDES.shiny}.)
}
\author{
Alex Chubaty
}
