% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonsUpdater.R
\name{polygonsUpdater}
\alias{polygonsUpdater}
\title{Polygons Updater Module}
\usage{
polygonsUpdater(input, output, session, proxy, polygonsInput, ...)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.}

\item{polygonsInput}{Reactive value which contains a set of polygons which should be displayed on the leaflet map}

\item{...}{Additional parameters of \code{leaflet::addPolygons} function. See \code{?leaflet::addPolygons} for reference.}
}
\description{
This creates a shiny module which uses leaflet proxy (see \code{?leafletProxy}) to update polygons displayed on a leaflet map.
}
\author{
Mateusz Wyszynski
}
