% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacyToS.R
\name{privacyStatementUI}
\alias{privacyStatementUI}
\alias{privacyStatement}
\title{Generate a page displaying your app's Privacy Statement}
\usage{
privacyStatementUI(id)

privacyStatement(input, output, session, file = "PRIVACY.md", status = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{file}{Path to a markdown file containing the Privacy Statement
(default: \code{"PRIVACY.md"}).}

\item{status}{The status (color) of the box in which the \code{file} is rendered.
See \code{\link[shinydashboard]{box}}.}
}
\description{
Renders a markdown file containing your app's Privacy Statement.
}
\author{
Alex Chubaty
}
