% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\docType{data}
\name{proj4stringLFLT}
\alias{proj4stringLFLT}
\title{Leaflet \code{proj.4} string}
\format{An object of class \code{character} of length 1.}
\usage{
proj4stringLFLT
}
\description{
This is the projection needed to plot on \pkg{leaflet} maps.
}
\keyword{datasets}
