% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastersOverTime.R
\name{rastersOverTimeUI}
\alias{rastersOverTimeUI}
\alias{rastersOverTime}
\title{Rasters-over-time shiny module}
\usage{
rastersOverTimeUI(id)

rastersOverTime(input, output, session, rasterList, polygonList,
  map = leaflet(), colorTable, histTitle = "", sliderTitle = "",
  mapTitle = "", nPolygons, nRasters, rasterStepSize = 10, sim = NULL,
  cachePath = getPaths()$cachePath, cacheNotOlderThan = Sys.time())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rasterList}{List of rasters to be displayed.}

\item{polygonList}{List with sets of polygons. Each such set can be displayed on a leaflet map.}

\item{map}{Leaflet map to show raster and polygons on.}

\item{colorTable}{File that contains color values for tiles (passed to \code{\link{gdal2Tiles}}).}

\item{histTitle}{Title to be shown above the histogram.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{mapTitle}{Title to be shown above the map.}

\item{nPolygons}{The number of available polygons.}

\item{nRasters}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}

\item{sim}{A SpaDES simulation object (\code{simList}).}

\item{cachePath}{Path to cache folder.}

\item{cacheNotOlderThan}{Load an artifact from cache only if it was created after notOlderThan.}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

None. Invoked for the side-effect of creating a shiny server part.
}
\description{
A shiny module showing the values from a raster in time on a predefined map.
Allows changing polygons and rasters via slider.
Additionally, a histogram summary for each raster choice is shown.
}
\author{
Damian Rodziewicz

Alex Chubaty
}
