% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastersOverTime.R
\name{rastersOverTime}
\alias{rastersOverTime}
\title{Rasters Over Time Shiny Module}
\usage{
rastersOverTime(input, output, session, rasters, polygonsList, colorTableFile,
  map = leaflet(), rasterStepSize = 10, cachePath = "cache",
  cacheNotOlderThan = Sys.time())
}
\arguments{
\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rasters}{Set of rasters to be displayed.}

\item{polygonsList}{List with sets of polygons. Each such set can be displayed on a leaflet map.}

\item{colorTableFile}{File that contains color values for tiles.}

\item{map}{Leaflet map to show raster and polygons on.}

\item{rasterStepSize}{.}

\item{cachePath}{Path to cache folder.}

\item{cacheNotOlderThan}{Load an artifact from cache only if it was created after notOlderThan.}
}
\value{
None. Invoked for the side-effect of creating a shiny server part.
}
\description{
Function \code{rastersOverTime} creates a shiny module server function
             which displays rasters changing in time on a predefined map.
             A slider, on which you can choose which raster should be currently displayed,
             is also created. Moreover, a histogram summary for each raster choice is shown.
}
\author{
Damian Rodziewicz
}
