% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastersOverTime.R
\name{rastersOverTimeUI}
\alias{rastersOverTimeUI}
\title{Rasters Over Time Shiny Module}
\usage{
rastersOverTimeUI(id, mapTitle, sliderTitle, histogramTitle, polygonsNumber,
  rastersNumber, rasterStepSize = 10)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{mapTitle}{Title to be shown above the map.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{histogramTitle}{Title to be shown above the histogram.}

\item{polygonsNumber}{The number of available polygons.}

\item{rastersNumber}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.
}
\description{
A shiny module showing the values from a raster in time.
             Allows changing polygons and rasters via slider.
}
\author{
Damian Rodziewicz
}
