% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterUpdater.R
\name{rasterUpdater}
\alias{rasterUpdater}
\title{Raster Updater Module}
\usage{
rasterUpdater(input, output, session, proxy, raster, group, ...)
}
\arguments{
\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.
When \code{NULL} (default) a basic leaflet map is created
and raster is displayed on this map.}

\item{raster}{Reactive value containing \code{RasterLayer} object.}

\item{group}{Group of rasters we want to update}

\item{...}{Additional parameters passed to \code{\link[leaflet]{addRasterImage}}
function.}
}
\value{
None. Invoked for the side-effect of sending information to leaflet proxy.
}
\description{
Creates a shiny module which uses \code{\link[leaflet]{leafletProxy}}
             to display \code{RasterLayer} object on a leaflet map. This is only
             suitable for small to medium sized rasters.
}
\author{
Mateusz Wyszynski
}
