% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{readConfig}
\alias{readConfig}
\title{Read an app's YAML config file}
\usage{
readConfig(file)
}
\arguments{
\item{file}{path to a YAML config file (\file{_config.yml})}
}
\value{
Invisibly return the config list.
}
\description{
Read an app's YAML config file
}
\author{
Alex Chubaty
}
