% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCallModuleDirective}
\alias{renderCallModuleDirective}
\title{Render a callModule directive.}
\usage{
renderCallModuleDirective(name, serverParameters)
}
\arguments{
\item{name}{Module name. It is equal to module id at this point.}

\item{serverParameters}{Server parameters used when calling the module.}
}
\value{
Rendered callModule directive.
}
\description{
Render a callModule directive.
}
\author{
Damian Rodziewicz
}
