% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCallModuleDirectives}
\alias{renderCallModuleDirectives}
\title{Render callModule directives for provided modules.}
\usage{
renderCallModuleDirectives(modules)
}
\arguments{
\item{modules}{Tibble with modules meta data. Tibble format: type, name, uiParameters, serverParameters.}
}
\value{
Rendered callModule directives.
}
\description{
Render callModule directives for provided modules.
}
\author{
Damian Rodziewicz
}
