% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCopyright}
\alias{renderCopyright}
\title{Render the copyright holder name of the app.}
\usage{
renderCopyright(copyright)
}
\arguments{
\item{copyright}{Name of the app copyright holder.}
}
\value{
Rendered app copyright holder name.
}
\description{
Render the copyright holder name of the app.
}
\author{
Alex Chubaty
}
