% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderExperimentDirectives}
\alias{renderExperimentDirectives}
\title{Render callModule directives for experiment module if provided.}
\usage{
renderExperimentDirectives(modules)
}
\arguments{
\item{modules}{Tibble with modules metadata. Tibble format: type, name, id, parameters.}
}
\value{
Rendered callModule directives.
}
\description{
Additonal customizations can be provided via two files: \file{pre_experiment.R}
and \file{post_experiment.R}.
}
\author{
Alex Chubaty
}
