% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderMenuItem}
\alias{renderMenuItem}
\title{Render a menu item.}
\usage{
renderMenuItem(tabName, menuItemName, icon)
}
\arguments{
\item{tabName}{Name of the tab that this item corresponds to.}

\item{menuItemName}{Name of the menu item to display.}

\item{icon}{Icon to display next to menu item name.}
}
\value{
Rendered menu item.
}
\description{
Render a menu item.
}
\author{
Damian Rodziewicz
}
