% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderMenuItems}
\alias{renderMenuItems}
\title{Render menu items for provided layout and available modules.}
\usage{
renderMenuItems(layout, modules)
}
\arguments{
\item{layout}{Tibble with layout metadata. Tibble format: tabName, menuItemName, icon, moduleName.}

\item{modules}{Tibble with modules metadata. Tibble format: type, name, uiParameters, serverParameters.}
}
\value{
Rendered menu items.
}
\description{
Render menu items for provided layout and available modules.
}
\author{
Damian Rodziewicz
}
