% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderSidebar}
\alias{renderSidebar}
\title{Render additional sidebar footer info.}
\usage{
renderSidebar(text)
}
\arguments{
\item{text}{Custom text to insert into app sidebar footer.}
}
\value{
Rendered additional sidebar footer info.
}
\description{
Render additional sidebar footer info.
}
\author{
Alex Chubaty
}
