% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderSimInitDirective}
\alias{renderSimInitDirective}
\title{Render callModule directives for simInit module if provided.}
\usage{
renderSimInitDirective(modules)
}
\arguments{
\item{modules}{Tibble with modules metadata. Tibble format: type, name, id, parameters.}
}
\value{
Rendered callModule directives.
}
\description{
Render callModule directives for simInit module if provided.
}
\author{
Alex Chubaty
}
