% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderSpadesShinyServer}
\alias{renderSpadesShinyServer}
\title{Render and save server.R file.}
\usage{
renderSpadesShinyServer(appDir, appMetadata)
}
\arguments{
\item{appDir}{The directory path to use for the new app.}

\item{appMetadata}{Application metadata.}
}
\value{
None. Invoked for the side-effect of writing rendered template to file.
}
\description{
Render and save server.R file.
}
\author{
Damian Rodziewicz
}
