% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderTemplate}
\alias{renderTemplate}
\title{Render a template using \pkg{whisker} package and write the result to a file.}
\usage{
renderTemplate(templatePath, data, path)
}
\arguments{
\item{templatePath}{Path to the template file.}

\item{data}{Named list or environment with variables that will be used during rendering.}

\item{path}{Path where the rendered file should be saved.}
}
\value{
None. Invoked for the side-effect of writing generated template to file.
}
\description{
Render a template using \pkg{whisker} package and write the result to a file.
}
\author{
Damian Rodziewicz
}
