% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderTitle}
\alias{renderTitle}
\title{Render the name of the app.}
\usage{
renderTitle(title)
}
\arguments{
\item{title}{Name of the app to be displayed in the top left (above the sidebar).}
}
\value{
Rendered app name.
}
\description{
Render the name of the app.
}
\author{
Alex Chubaty
}
