% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonList.R
\name{rmNulls}
\alias{rmNulls}
\title{Remove elements of a list whose subelements contain a NULL value}
\usage{
rmNulls(x, ...)
}
\arguments{
\item{x}{A named list.}

\item{...}{Additonal arguments for specific methods.}
}
\description{
Remove elements of a list whose subelements contain a NULL value
}
\author{
Alex Chubaty
}
