% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{shinyModule}
\alias{shinyModule}
\title{Use an existing shiny module.}
\usage{
shinyModule(moduleName)
}
\arguments{
\item{moduleName}{Name of the module to use.}
}
\value{
Metadata object that describes an existing shiny module.
}
\description{
This method creates a metadata object to use in application metadata.
Created object describes an existing shiny module.
}
\author{
Damian Rodziewicz
}
