% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{simEventDiagramUI}
\alias{simEventDiagramUI}
\alias{simEventDiagram}
\title{Simulation event diagram shiny module}
\usage{
simEventDiagramUI(id)

simEventDiagram(input, output, session, sim)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
Simulation event diagram shiny module
}
\author{
Alex Chubaty and Greyson Wang (module)
}
