% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{simInfoUI}
\alias{simInfoUI}
\alias{simInfo}
\title{\code{SpaDES} simulation information (parent) module}
\usage{
simInfoUI(id)

simInfo(input, output, session, sim, elements = NULL)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}

\item{elements}{Character vector describing which simulation diagrams to include as tabs.
Default \code{NULL}, for all diagrams (i.e., \code{c("modules", "objects", "events")}).}
}
\description{
\code{SpaDES} simulation information (parent) module
}
\author{
Olivia Sung, Alex Chubaty, Greyson Wang
}
