% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{simLogFile}
\alias{simLogFile}
\title{Get simulation-specific log filepath}
\usage{
simLogFile(config, session, simID)
}
\arguments{
\item{config}{\pkg{SpaDES.shiny} app configuration object (e.g., from \code{\link{readConfig}})}

\item{session}{shiny server session object}

\item{simID}{The simulation ID}
}
\description{
Get simulation-specific log filepath
}
\author{
Alex Chubaty
}
