% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{simModuleDiagramUI}
\alias{simModuleDiagramUI}
\alias{simModuleDiagram}
\title{\code{SpaDES} module diagram module}
\usage{
simModuleDiagramUI(id)

simModuleDiagram(input, output, session, sim)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
\code{SpaDES} module diagram module
}
\author{
Alex Chubaty and Greyson Wang (module)
}
