% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOutputs.R
\name{simOutputsUI}
\alias{simOutputsUI}
\alias{simOutputs}
\title{Simulation outputs (graphs and figures) module}
\usage{
simOutputsUI(id)

simOutputs(input, output, session, sim)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
TO DO: needs description
}
\author{
Alex Chubaty
}
