% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider}
\alias{slider}
\title{Slider Module}
\usage{
slider(input, output, session)
}
\arguments{
\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}
}
\description{
Slider Module
}
