% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{sliderUI}
\alias{sliderUI}
\title{Slider Module UI}
\usage{
sliderUI(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{...}{sliderInput arguments. See ?sliderInput for reference.}
}
\description{
This creates a slider module UI.
}
\author{
Mateusz Wyszynski
}
