% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-shiny-package.R
\docType{package}
\name{SpaDES.shiny-package}
\alias{SpaDES.shiny}
\alias{SpaDES.shiny-package}
\title{The \pkg{SpaDES.shiny} package}
\description{
The \pkg{SpaDES.shiny} package
}
\seealso{
Useful links:
\itemize{
  \item \url{http://spades-shiny.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.shiny}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.shiny/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alex.chubaty@gmail.com}

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@canada.ca}
  \item Damian Rodziewicz
  \item Mateusz Wyszynski
}

Other contributors:
\itemize{
  \item Olivia Sung [contributor]
  \item Alex Tso [contributor]
  \item Greyson Wang [contributor]
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
