% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades.R
\name{spades_expt}
\alias{spades_expt}
\title{Shiny module to run a \code{SpaDES} simulation experiment}
\usage{
spades_expt(
  input,
  output,
  session,
  sim,
  reps,
  seed,
  objectsToHash,
  cacheDebug = "complete",
  spadesDebug = getOption("spades.debug"),
  ...
)
}
\arguments{
\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A reactive \code{simList} object (e.g., pass \code{sim = sim()}).}

\item{reps}{A reactive indicating the number of replicates (e.g., pass \code{reps = reps()}).}

\item{seed}{An integer to pass to \code{set.seed} for the simulation experiment.}

\item{objectsToHash}{A list of objects to hash (with \code{Cache}).}

\item{cacheDebug}{\code{Cache()} debugging (default \code{"complete"}). See \code{\link{Cache}}.}

\item{spadesDebug}{\code{spades()} debugging (default \code{getOption("spades.debug")}).
See \code{link{spades}}.}

\item{...}{Additional arguments passed to \code{experiment}.}
}
\value{
A list of \code{simList} objects.
}
\description{
Simply a wrapper around \code{\link[SpaDES.core]{experiment}} to be called as a shiny module.
}
\note{
This is a server-only module with no UI component.
}
\seealso{
\code{\link[SpaDES.core]{experiment}}
}
\author{
Alex Chubaty
}
