% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades.R
\name{spades_simInit}
\alias{spades_simInit}
\title{Shiny module to initialize a \code{SpaDES} simulation}
\usage{
spades_simInit(input, output, session, ...)
}
\arguments{
\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{...}{Additional arguments passed to \code{simInit}.}
}
\value{
A reactive \code{simList} object
}
\description{
Simply a wrapper around \code{\link[SpaDES.core]{simInit}} to be called as a shiny module.
}
\note{
This is a server-only module with no UI component.
}
\seealso{
\code{\link[SpaDES.core]{simInit}}
}
\author{
Alex Chubaty
}
