% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPopups.R
\name{summaryPopups}
\alias{summaryPopups}
\title{Summary popups (shiny module)}
\usage{
summaryPopups(input, output, session, proxy, click, rast, poly,
  rasterValueLabel = "Raster value: \%s", extractedValues = NULL)
}
\arguments{
\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.}

\item{click}{Reactive value with click on shape input from leaflet map.}

\item{rast}{Reactive value with raster to summarize by.}

\item{poly}{Reactive value with current polygon on the map.}

\item{rasterValueLabel}{String with description of raster value. Uses \code{sprintf},
so must include exactly one \code{\%s} in order to display raster value.
Default is a string \code{"Raster value: \%s"}.}

\item{extractedValues}{List of attributes from \code{SpatialPolygonDataFrame}
which should be included in popup summary.
When \code{NULL} (default) all attributes are included.}
}
\value{
None. Invoked for the side-effect of creating a shiny observer.
}
\description{
Add popups with polygon and raster summaries on a leaflet map.
}
\note{
This is a server-only module with no UI component.
}
\author{
Mateusz Wyszynski
}
