% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacyToS.R
\name{termsOfServiceUI}
\alias{termsOfServiceUI}
\alias{termsOfService}
\title{Generate a page displaying your app's Terms of Service}
\usage{
termsOfServiceUI(id)

termsOfService(input, output, session, file = "TERMS.md", status = NULL)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{file}{Path to a markdown file containing the Terms of Service
(default: \code{"TERMS.md"})}

\item{status}{The status (colour) of the box in which the \code{file} is rendered.
See \code{\link[shinydashboard]{box}}.}
}
\description{
Renders a markdown file containing your app's Terms of Service.
}
\author{
Alex Chubaty
}
