% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesUpdater.R
\name{tilesUpdater}
\alias{tilesUpdater}
\title{tilesUpdater Module}
\usage{
tilesUpdater(input, output, session, proxy, urlTemplate, tilesGroup,
  addTilesParameters = NULL, addLayersControlParameters = NULL)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.}

\item{urlTemplate}{Reactive value which contains an url with the tiles that should be added to the leaflet map as a result.}

\item{tilesGroup}{Group of rasters/tiles we want to update}

\item{addTilesParameters}{Additional parameters of \code{leaflet::addTiles} function. See \code{?addTiles} for reference. Parameters: \code{proxy}/\code{map}, code{urlTemplate} and \code{group} should be provided directly as \code{tilesUpdater} arguments using \code{proxy}, \code{urlTemplate} and \code{tilesGroup} parameter respectively. If passed within \code{addTilesParameters}, these arguments will be surpressed by the values passed directly to \code{tilesUpdater}.}

\item{addLayersControlParameters}{Additional parameters of \code{leaflet::addLayersControl} function. See \code{?addLayersControl} for reference.}
}
\description{
This creates a shiny module which uses leaflet proxy (see \code{?leafletProxy}) to update tiles displayed on a leaflet map.
}
\author{
Mateusz Wyszynski
}
