% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesofRasters.R
\name{timeSeriesofRastersUI}
\alias{timeSeriesofRastersUI}
\alias{timeSeriesofRasters}
\title{Time series of rasters shiny module}
\usage{
timeSeriesofRastersUI(id)

timeSeriesofRasters(input, output, session, rctRasterList, rctUrlTemplate,
  rctPolygonList, rctChosenPoly, defaultPolyName = NULL,
  shpStudyRegionName = NULL, shpStudyRegionLFLT = NULL, colorPalette,
  maxAge, zoom = 5, mapTilesDir = "www/", mapLegend = "",
  mapTitle = "", sliderTitle = "", histTitle = "", nPolygons,
  nRasters, rasterStepSize = 10, uploadOpts = list(auth = NULL, path =
  NULL, user = NULL), rctStudyArea = NULL, thinKeep = 0.05)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{rctRasterList}{A reactive that gives a list of rasters to be displayed.}

\item{rctUrlTemplate}{The reactive url template for leaflet map tiles}

\item{rctPolygonList}{Reactive list with sets of polygons to be displayed on a leaflet map.
# TODO: describe the format of the list!}

\item{rctChosenPoly}{The user-selected polygon.}

\item{defaultPolyName}{Name of the polygon to use as the default for mapping.}

\item{shpStudyRegionName}{Name of the study area region (from \code{rctRasterList}).}

\item{shpStudyRegionLFLT}{\code{SpatialPolygonDataFrame} for the study region.}

\item{colorPalette}{Colour palette to use.}

\item{maxAge}{Maximum simulation age.}

\item{zoom}{Initial leaflet zoom.}

\item{mapTilesDir}{Directory where pre-generated map tiles are stored
(default \code{"www/"}; note the trailing slash).
Be sure to add this dir as 'tiles' using \code{shiny::addResourcePath}
in your \file{global.R}.
E.g., \code{shiny::addResourcePath("tiles", "www/All/FULL/map-tiles")}
(note we don't need the trailing slash here).}

\item{mapLegend}{The legend text to add to the leaflet map.}

\item{mapTitle}{Title to be shown above the map.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{histTitle}{Title to be shown above the histogram.}

\item{nPolygons}{The number of available polygons.}

\item{nRasters}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}

\item{uploadOpts}{A list of options for use with file uploads:
\code{auth} logical indicating whether user is authorized to upload;
\code{path} a directory path to use for file uploads;
\code{user} the current username (used for creating user-specific paths).
The default for all options is \code{NULL}, which means do not use.}

\item{rctStudyArea}{A reactive \code{SpatialPolygons*} object for the whole study area region.}

\item{thinKeep}{Proportion of points in polygons to keep when thinning or plotting.
See \code{\link[rmapshaper]{ms_simplify}}.}
}
\value{
Invoked for the side-effect of creating a shiny UI.

Reactive polygon selected by the user with the \code{polygonChooser} module.
         Invoked for the side-effect of creating shiny server and UI components. # TODO: reword
}
\description{
A shiny module that displays rasters changing in time on a predefined map,
with the user choosing which raster should be displayed by using the slider.
A histogram summary for each raster choice is also shown.
}
\author{
Mateusz Wyszynski

Alex Chubaty
}
