% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSinceFire.R
\name{timeSinceFireUI}
\alias{timeSinceFireUI}
\alias{timeSinceFire}
\title{Time Since Fire Shiny Module UI}
\usage{
timeSinceFireUI(id, rastersNumber)

timeSinceFire(input, output, session, rasters, leafletZoomInit = 5,
  studyArea = "SMALL")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function}

\item{rastersNumber}{How many rasters can we choose from}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{rasters}{Set of rasters to be displayed}

\item{leafletZoomInit}{Initial leaflet zoom}

\item{studyArea}{Size of study area. Options: "FULL", "EXTRALARGE", "LARGE", "MEDIUM", "NWT", "SMALL"}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

None. Invoked for the side-effect of creating a shiny server part.
}
\description{
Function \code{timeSinceFireUI} creates a shiny module UI for Time Since Fire shiny module.

Function \code{timeSinceFire} creates a shiny module server function which displays rasters changing in time on a predefined map.
A slider on which you can choose which raster should be currently displayed is also created.
Moreover, a histogram summary for each raster choice is shown.
}
\author{
Mateusz Wyszynski
}
