% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSinceFire.R
\name{timeSinceFireUI}
\alias{timeSinceFireUI}
\title{Time Since Fire Shiny Module UI}
\usage{
timeSinceFireUI(id, rastersNumber, polygonsNumber)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function}

\item{rastersNumber}{How many rasters can we choose from}
}
\description{
This creates a shiny module UI for Time Since Fire shiny module. Compare with \code{?timeSinceFire}
}
\author{
Mateusz Wyszynski
}
