% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{updateConfig}
\alias{updateConfig}
\title{Update yaml config file}
\usage{
updateConfig(fileFrom, fileTo)
}
\arguments{
\item{fileFrom}{Path to yaml config file to update from}

\item{fileTo}{Path to yaml config file}
}
\description{
E.g., copy updated config from git repo to app config without clobbering
}
\author{
Alex Chubaty
}
