% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonList.R
\docType{methods}
\name{updateList,polygonList,polygonList-method}
\alias{updateList,polygonList,polygonList-method}
\title{Update polygon list}
\usage{
\S4method{updateList}{polygonList,polygonList}(x, y)
}
\arguments{
\item{x}{a named list}

\item{y}{a named list}
}
\value{
A named list, with elements sorted by name.
         The values of matching elements in list \code{y}
         replace the values in list \code{x}.
}
\description{
Update polygon list
}
\note{
This is a temporary workaround until we resolve inheritance of S3 classes (see \code{.polygonList}).
}
\author{
Alex Chubaty
}
