% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadPolygon.R
\name{uploadPolygonUI}
\alias{uploadPolygonUI}
\alias{uploadPolygon}
\title{Upload a shapefile}
\usage{
uploadPolygonUI(id)

uploadPolygon(input, output, session, authStatus, userDir, studyArea)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{authStatus}{Logical indicating whether a user can upload files.}

\item{userDir}{User-specific directory in which to store uploaded files.}

\item{studyArea}{A \code{Spatial} object used as a template for postprocessing
the uploaded polygon, which is cropped, reprojected, etc.
to match \code{studyArea}. See \code{\link[SpaDES.tools]{postProcess}}.}
}
\value{
Reactive object containing the uploaded polygon.
}
\description{
Upload a shapefile
}
\author{
Alex Chubaty
}
