% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeRaster.R
\name{visualizeRaster}
\alias{visualizeRaster}
\title{Visualize Raster Module Server Function}
\usage{
visualizeRaster(input, output, session, raster, ...)
}
\arguments{
\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{raster}{Raster which should be visualized. Might be a reactive value
containing raster.}

\item{...}{Additional parameters passed to \code{link[rasterVis]{levelplot}}.}
}
\value{
None. Invoked for the side-effect of rendering bar plot.
}
\description{
Visualize Raster Module Server Function
}
