% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeRaster.R
\name{visualizeRasterUI}
\alias{visualizeRasterUI}
\title{Visualize Raster Module UI Function}
\usage{
visualizeRasterUI(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{Additional parameters passed to \code{link[shiny]{plotOutput}}.}
}
\value{
None. Invoked for the side-effect of generating UI for plot.
}
\description{
Shiny module which visualizes raster.
}
