% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authGoogle.R
\name{authGoogleUI}
\alias{authGoogleUI}
\alias{authGoogle}
\title{Authorize a user using Google Drive account}
\usage{
authGoogleUI(id)

authGoogle(input, output, session, authFile)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{authFile}{Private url to a Google Drive file with access control.}
}
\description{
In Google Drive, create a blank text file (e.g., name_of_your_app.txt) with
access control settings that allow only authorized users to download/view etc.
}
\details{
Copy the (private) Drive file url and pass that as the \code{authFile} argument.

Authentication status is stored in \code{session$userData$AUTH_GOOGLE()}, a reactive value.
}
\author{
Alex Chubaty
}
