% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authGoogle.R
\name{authGoogleUI}
\alias{authGoogleUI}
\alias{authGoogle}
\title{Authorize a user using Google Drive account}
\usage{
authGoogleUI(id)

authGoogle(input, output, session, appURL, authFile, icon = "google")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{appURL}{URL to the hosted app.}

\item{authFile}{Private url to a Google Drive file with access control.}

\item{icon}{Default \code{"google"}. Name of icon to display beside the login button.
Use of an icon with \pkg{shinydashboard} may produce undesired results.
Disable use of the icon with \code{icon = NULL}.}
}
\value{
A list of the user's Google profile details (name, email, etc.)
}
\description{
Logs a user into their Google account for authorized use of the app.
Allows app content to be made conditionally available to 3 classes of user:
}
\details{
1. Anonymous (non-logged in) users;
2. Logged in users who are not autheticated against a whitelist;
3. Logged in and authenticated users.

This mechanism relies on the file permissions set on a file in Google Drive.
Simply create a blank text file (e.g., name_of_your_app.txt) in Google Drive
and edit the access control settings to allow only authorized users to download/view
this file.
You will need the private URL (link) to this file.
Assign this URL to a variable in your \file{global.R} and pass that as the
\code{authFile} argument when calling the \code{authGoogle} module.

This allows access control to be managed via Google rather than locally on the shiny server.

Login status is stored in \code{session$userData$userLoggedIn()}, a reactive value.
Authentication status is stored in \code{session$userData$userAuthorized()}, a reactive value.
}
\note{
Based on \code{googleAuthR::googleAuth} and \code{googleAuthR::googleAuthUI}.
}
\section{Additional requirements}{

Your \file{global.R} file should set the following options:
1. `googleAuthR.scopes.selected`: `c("https://www.googleapis.com/auth/drive.readonly", "https://www.googleapis.com/auth/userinfo.email", "https://www.googleapis.com/auth/userinfo.profile")`.
2. `googleAuthR.webapp.client_id`: your Google app oauth id.
3. `googleAuthR.webapp.client_secret`: your Google app oauth "secret".

Be sure to also set \code{appURL} and \code{authFile} in \file{global.R}.

See the authentication vignette (\code{vignette("authentication", "SpaDES.shiny")}).
}

\author{
Alex Chubaty

Mark Edmondson
}
