% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{writeConfig}
\alias{writeConfig}
\title{Write YAML config file}
\usage{
writeConfig(config, file)
}
\arguments{
\item{config}{List of configuration variables to write to file}

\item{file}{Path to YAML config file}
}
\description{
Write YAML config file
}
