% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyrightAuthors.R
\name{copyrightAuthorsUI}
\alias{copyrightAuthorsUI}
\alias{copyrightAuthors}
\title{Generate Authors and Copyright Information}
\usage{
copyrightAuthorsUI(id)

copyrightAuthors(input, output, session, appName, authorInfo, copyrightInfo,
  licenseFile = "LICENSE", status = "primary")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{appName}{Character string giving the name of your SpaDES.shiny app.}

\item{authorInfo}{Character string or person list containing your app's
authorship info.}

\item{copyrightInfo}{Character string containing your app's copyright info.}

\item{licenseFile}{Filepath to your app's LICENSE file
(default is \file{LICENSE} in your shiny app dir).}

\item{status}{The boxes' status, passed to \code{\link[shinydashboard]{box}}
(default \code{"primary"}, which results in dark blue).}
}
\description{
Generate Authors and Copyright Information
}
\details{
Generates a fluidRow containing two boxes
about the app's Authors and Contributors and the Copyright and License.
}
\author{
Alex Chubaty & Alex Tso (module)
}
