% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footers.R
\name{copyrightFooterUI}
\alias{copyrightFooterUI}
\alias{copyrightFooter}
\alias{sidebarFooterUI}
\alias{sidebarFooter}
\title{Generate HTML footers for body or sidebar}
\usage{
copyrightFooterUI(id)

copyrightFooter(
  input,
  output,
  session,
  cph = "Author Name",
  year = format(Sys.time(), "\%Y")
)

sidebarFooterUI(id)

sidebarFooter(input, output, session, footer = NULL)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{cph}{Character string indicating the copyright holder name.
This is automatically generated from the \code{copyright} item
in the app metadata sued my \code{newApp}.}

\item{year}{Character string indicating the copyright date.
Defaults to the current year.}

\item{footer}{Custom text to appear in footer above the 'SpaDES' and 'Appsilon' logos.}
}
\description{
App copyright info can be inserted into a footer in the dashboard body using
\code{copyrightFooter} module.
}
\author{
Alex Chubaty
}
