% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footers.R
\name{copyrightFooter}
\alias{copyrightFooter}
\alias{sidebarFooter}
\title{Generate HTML footers for body or sidebar}
\usage{
copyrightFooter(copyrightInfo)

sidebarFooter()
}
\arguments{
\item{copyrightInfo}{Character string containing your app's copyright info.}
}
\description{
App copyright info can be inserted into a footer in the dashboard body using
\code{copyrightFooter}.
}
\examples{
copyrightInfo <- paste(
  shiny::icon("copyright",  lib = "font-awesome"), "Copyright ",
  format(Sys.time(), "\%Y"),
  paste("Her Majesty the Queen in Right of Canada,",
        "as represented by the Minister of Natural Resources Canada.")
)

# The following line is in the ui.R.template and will be run if 
# global.R has a copyrightInfo object
copyrightFooter(copyrightInfo)
}
\author{
Alex Chubaty

Alex Chubaty
}
