% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{gdal2Tiles}
\alias{gdal2Tiles}
\title{Convert a raster to tile images}
\usage{
gdal2Tiles(raster, outputPath, zoomRange, colorTableFile)
}
\arguments{
\item{raster}{Raster to generate tiles from.}

\item{outputPath}{Output folder path.}

\item{zoomRange}{List of numbers representing zoom value to generate tiles for.}

\item{colorTableFile}{Table file that contains colour values for tiles.}
}
\value{
Leaflet map url template path. # TODO: improve description
}
\description{
Important: \code{gdalUtils_gdalPath} environment variable must point to
             path where GDAL is installed.
}
\note{
Requires GDAL (>=2) and python are found on the system \code{PATH}.
}
\author{
Damian Rodziewicz

Alex Chubaty
}
