% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal2Tiles.R
\name{gdal2Tiles}
\alias{gdal2Tiles}
\title{Gdal2Tiles}
\usage{
gdal2Tiles(raster, outputPath, zoomRange, colorTextFile)
}
\arguments{
\item{raster}{Raster to generate tiles from.}

\item{outputPath}{Output folder path.}

\item{zoomRange}{List of numbers representing zoom value to generate tiles for.}

\item{colorTextFile}{Text file that contains color values for tiles.}
}
\value{
None. Invoked for the side-effect of creating tile images.
}
\description{
Function that converts given raster to tile images.
             Important: gdalUtils_gdalPath environment variable must point to
             path where gdal is installed.
}
\author{
Damian Rodziewicz
}
