% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{gdal2Tiles}
\alias{gdal2Tiles}
\title{Convert a raster to tile images}
\usage{
gdal2Tiles(raster, outputPath, zoomRange, colorTableFile)
}
\arguments{
\item{raster}{Raster to generate tiles from.}

\item{outputPath}{Output folder path.}

\item{zoomRange}{List of numbers representing zoom value to generate tiles for.}

\item{colorTableFile}{Table file that contains color values for tiles.}
}
\value{
None. Invoked for the side-effect of creating tile images.
}
\description{
Important: `gdalUtils_gdalPath` environment variable must point to
             path where GDAL is installed.
}
\note{
Requires GDAL (>=2) and python are found on the system PATH.
}
\author{
Damian Rodziewicz
}
