% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{getModuleById}
\alias{getModuleById}
\title{Retrieve a module metadata from modules tibble.}
\usage{
getModuleById(modules, moduleId)
}
\arguments{
\item{modules}{\code{tibble} with modules metadata.
\code{tibble} format: \code{type}, \code{name}, \code{id}, \code{parameters}.}

\item{moduleId}{id of the module to retrieve.}
}
\value{
\code{tibble} containing the module if it was found. Empty \code{tibble} otherwise.
}
\description{
Retrieve a module metadata from modules tibble.
}
\author{
Damian Rodziewicz
}
