% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{.getSubtable}
\alias{.getSubtable}
\title{Get subtable from a \code{data.table}}
\usage{
.getSubtable(dataTable, chosenCategories, chosenValues)
}
\arguments{
\item{dataTable}{A \code{data.table}}

\item{chosenCategories}{...}

\item{chosenValues}{...}
}
\description{
Get subtable from a \code{data.table}
}
