% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramUI}
\alias{histogramUI}
\alias{histogram}
\title{Histogram Module}
\usage{
histogramUI(id, ...)

histogram(input, output, session, data, globalData, chosenCategories,
  numberOfSimulationTimes)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{...}{Additional parameters passed to \code{link[shiny]{plotOutput}}}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{data}{Subset of \code{globalData} created during slicing.
See \code{?slicer} for reference.}

\item{globalData}{Data in form of data table. This data was sliced during slicing.
See \code{?slicer} for reference.}

\item{chosenCategories}{List with categories that where chosen when slicing the data.}

\item{numberOfSimulationTimes}{How many simulation times occurred in simulation.}
}
\value{
None. Invoked for the side-effect of generating UI for plot

None. Invoked for the side-effect of rendering bar plot.
}
\description{
Shiny module which creates a histogram using barplot based on the data received
}
