% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram Module Server Function}
\usage{
histogram(input, output, session, data, addAxisParams = NULL, ...)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{data}{Reactive value containing a data table. This is global data.
Desired subtables can be retrieved using \code{chosenCategories}
and \code{chosenValues} parameters.}

\item{addAxisParams}{Reactive value with parameters to \code{\link[graphics]{axis}}.
If \code{NULL} (default) then no axis is drawn.}

\item{...}{Additional arguments passed to \code{\link[graphics]{barplot}} function.}
}
\value{
None. Invoked for the side-effect of rendering bar plot.
}
\description{
Server function of a shiny module which creates a histogram
             using barplot based on the data received.
}
