% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramUI}
\alias{histogramUI}
\alias{histogram}
\title{Histogram Module UI Function}
\usage{
histogramUI(id, ...)

histogram(input, output, session, histdata, addAxisParams = NULL,
  verticalBar = NULL, fname = NULL, ...)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{...}{Additional UI parameters passed to \code{link[shiny]{plotOutput}},
or additional server arguments passed to \code{\link[graphics]{barplot}}.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{histdata}{Reactive value containing a numeric vector of proportions corresponding
to each histogram bin (i.e., the output of \code{hist(..., plot = FALSE)}
represented as a proportion).
Desired subtables can be retrieved using \code{chosenCategories}
and \code{chosenValues} parameters.}

\item{addAxisParams}{Reactive value with parameters to \code{\link[graphics]{axis}}.
If \code{NULL} (default) then no axis is drawn.}

\item{verticalBar}{Numeric value at which to add an \code{abline} to the histogram.}

\item{fname}{optional filepath to save \code{.png} outputs}
}
\value{
None. Invoked for the side-effect of rendering bar plot.
}
\description{
UI function of a shiny module which creates a histogram
             using barplot based on the data received

Server function of a shiny module which creates a histogram
             using barplot based on the data received.
}
