% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramUI}
\alias{histogramUI}
\alias{histogram}
\title{Histogram Module UI Function}
\usage{
histogramUI(id, ...)

histogram(input, output, session, data, addAxisParams = NULL, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{...}{Additional UI parameters passed to \code{link[shiny]{plotOutput}},
or additional server arguments passed to \code{\link[graphics]{barplot}}.}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{data}{Reactive value containing a data table. This is global data.
Desired subtables can be retrieved using \code{chosenCategories}
and \code{chosenValues} parameters.}

\item{addAxisParams}{Reactive value with parameters to \code{\link[graphics]{axis}}.
If \code{NULL} (default) then no axis is drawn.}
}
\value{
None. Invoked for the side-effect of rendering bar plot.
}
\description{
UI function of a shiny module which creates a histogram
             using barplot based on the data received

Server function of a shiny module which creates a histogram
             using barplot based on the data received.
}
