% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramForRaster.R
\name{histogramForRasterUI}
\alias{histogramForRasterUI}
\alias{histogramForRaster}
\title{Histogram for Raster Shiny Module}
\usage{
histogramForRasterUI(id)

histogramForRaster(input, output, session, raster)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{raster}{Reactive value containing raster}
}
\description{
This creates UI for shiny module creating histogram of data contained in raste


}
