% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramForRaster.R
\name{histogramForRasterUI}
\alias{histogramForRasterUI}
\alias{histogramForRaster}
\title{Histogram for Raster Shiny Module}
\usage{
histogramForRasterUI(id, title = "", plotParameters, ...)

histogramForRaster(
  input,
  output,
  session,
  rctRasterVals,
  rctHistogramBreaks,
  scale = 1,
  addAxisParams = NULL,
  ...
)
}
\arguments{
\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{title}{Optional title for the histogram. Any shiny tag can be used.}

\item{plotParameters}{A list of parameters passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}

\item{...}{Additional parameters passed to \code{\link[shinydashboard:box]{shinydashboard::box()}}
in \code{histogramForRasterUI}, and \code{\link[graphics:barplot]{graphics::barplot()}}
in \code{histogramForRaster}.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{rctRasterVals}{Reactive values (likely extracted from a raster)}

\item{rctHistogramBreaks}{Reactive value which is responsible for \code{breaks}
parameter, as in \code{\link[graphics:hist]{graphics::hist()}}.}

\item{scale}{Number used for scaling heights of histogram bars.
When set to 1 (default) histogram bar height represents the number
of raster cells with value from bar interval.
If the resolution of raster is known, \code{scale} can be used
to transform these heights into the ones representing area covered
by cells (count).
When set to 2 (or, more generally, some number n with no further meaning)
this will just increase the height of each histogram bar by 2
(n, respectively).
So, in this scenario, each histogram's bar height is just count
times 2 (count times n).}

\item{addAxisParams}{Reactive value with parameters to \code{\link[graphics:axis]{graphics::axis()}}.
If \code{NULL} (default) then no axis is drawn.}
}
\value{
None. Invoked for the side-effect of creating shiny UI.

None. Invoked for the side-effect of rendering histogram plot.
}
\description{
A shiny module creating histogram of data contained in a raster.
}
