% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{moduleInfoUI}
\alias{moduleInfoUI}
\alias{moduleInfo}
\title{Detailed \code{SpaDES} module info module}
\usage{
moduleInfoUI(id)

moduleInfo(input, output, session, sim, data = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object.}

\item{data}{Logical. \code{TRUE} indicates to use the Rmd file for the
module's data (\file{moduleName/moduleName_data.Rmd}), if present.
\code{FALSE} uses the module's Rmd file (\file{moduleName/moduleName.Rmd}).}
}
\description{
TO DO: needs documentation
}
\author{
Alex Chubaty
}
