% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{newApp}
\alias{newApp}
\title{Create a new \pkg{SpaDES.shiny} app}
\usage{
newApp(appDir, appMetadata)
}
\arguments{
\item{appDir}{The directory path to use for the new app.}

\item{appMetadata}{Application metadata.}
}
\value{
None. Invoked for the side-effect of creating a new app.
}
\description{
Setup the necessary directory structure for a new app, and create a new app
based on a template.
}
\author{
Damian Rodziewicz
}
