% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonsUpdater.R
\name{polygonsUpdater}
\alias{polygonsUpdater}
\title{Polygons updater (shiny module)}
\usage{
polygonsUpdater(input, output, session, proxy, poly, group = "group", ...)
}
\arguments{
\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.}

\item{poly}{A reactive containing a polygon to be displayed on the leaflet map.}

\item{group}{Group of polygons we want to update.}

\item{...}{Additional parameters of \code{\link[leaflet]{addPolygons}} function.}
}
\value{
None. Invoked for the side-effect of creating an observer, which
        manages current set of polygons on a leaflet map using leaflet proxy.
}
\description{
Update polygons displayed on a leaflet map using \code{\link[leaflet]{leafletProxy}}.
}
\note{
This is a server-only module with no UI component.
}
\author{
Mateusz Wyszynski
}
