% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonsUpdater.R
\name{polygonsUpdater}
\alias{polygonsUpdater}
\title{Polygons Updater Module}
\usage{
polygonsUpdater(input, output, session, proxy, polygons, ...)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{proxy}{Leaflet proxy which manages a connected leaflet map.}

\item{polygons}{Reactive value which contains a set of polygons which should be displayed on the leaflet map}

\item{...}{Additional parameters of \code{\link[leaflet]{addPolygons}} function.}
}
\value{
None. Invoked for the side-effect of creating an observer
        which manages current set of polygons on a leaflet map using leaflet proxy.
}
\description{
This creates a shiny module which uses leaflet proxy (see \code{\link[leaflet]{leafletProxy}})
             to update polygons displayed on a leaflet map.
}
\author{
Mateusz Wyszynski
}
