% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastersOverTime.R
\name{rastersOverTimeUI}
\alias{rastersOverTimeUI}
\alias{rastersOverTime}
\title{Rasters Over Time Shiny Module}
\usage{
rastersOverTimeUI(id, mapTitle, sliderTitle, histogramTitle, polygonsNumber,
  rastersNumber, rasterStepSize = 10)

rastersOverTime(input, output, session, rasters, polygonsList, colorTableFile,
  map = leaflet(), rasterStepSize = 10, sim = NULL,
  cachePath = getPaths()$cachePath, cacheNotOlderThan = Sys.time())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{mapTitle}{Title to be shown above the map.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{histogramTitle}{Title to be shown above the histogram.}

\item{polygonsNumber}{The number of available polygons.}

\item{rastersNumber}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rasters}{Set of rasters to be displayed.}

\item{polygonsList}{List with sets of polygons. Each such set can be displayed on a leaflet map.}

\item{colorTableFile}{File that contains color values for tiles.}

\item{map}{Leaflet map to show raster and polygons on.}

\item{sim}{A SpaDES simulation object (\code{simList}).}

\item{cachePath}{Path to cache folder.}

\item{cacheNotOlderThan}{Load an artifact from cache only if it was created after notOlderThan.}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

None. Invoked for the side-effect of creating a shiny server part.
}
\description{
A shiny module showing the values from a raster in time on a predefined map.
             Allows changing polygons and rasters via slider.
             Additionally, a histogram summary for each raster choice is shown.
}
\author{
Damian Rodziewicz

Damian Rodziewicz
}
