% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastersOverTime.R
\name{rastersOverTimeUI}
\alias{rastersOverTimeUI}
\alias{rastersOverTime}
\title{Rasters-over-time shiny module}
\usage{
rastersOverTimeUI(id)

rastersOverTime(input, output, session, rctRasterList, rctUrlTemplate,
  rctPolygonList, rctChosenPolyName, defaultPolyName = NULL,
  map = leaflet(), mapTilesDir = "www/", colorPalette,
  histTitle = "", sliderTitle = "", mapTitle = "", nPolygons,
  nRasters, rasterStepSize = 10)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rctRasterList}{A reactive that gives a list of rasters to be displayed.}

\item{rctUrlTemplate}{The reactive url template for leaflet map tiles}

\item{rctPolygonList}{Reactive list with sets of polygons to be displayed on a leaflet map.
# TODO: decribe the format of the list!}

\item{rctChosenPolyName}{Reactive containing the name of the selected polygon (character).}

\item{defaultPolyName}{The name fo the default selected polygon.}

\item{map}{Leaflet map to show raster and polygons on.}

\item{mapTilesDir}{Directory where pre-generated map tiles are stored
(default \code{"www/"}; note the trailing slash).
Be sure to add this dir as 'tiles' using \code{shiny::addResourcePath}
in your \file{global.R}.
E.g., \code{shiny::addResourcePath("tiles", "www/All/FULL/map-tiles")}
(note we don't need the trailing slash here).}

\item{colorPalette}{Colour palette to use.}

\item{histTitle}{Title to be shown above the histogram.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{mapTitle}{Title to be shown above the map.}

\item{nPolygons}{The number of available polygons.}

\item{nRasters}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

Reactive polygon selected by the user with the \code{polygonChooser} module.
         Invoked for the side-effect of creating shiny server and ui components. # TODO: reword
}
\description{
A shiny module showing the values from a raster in time on a predefined map.
Allows changing polygons and rasters via slider.
Additionally, a histogram summary for each raster choice is shown.
}
\author{
Damian Rodziewicz

Alex Chubaty
}
