% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCallModuleDirective}
\alias{renderCallModuleDirective}
\title{Render a \code{callModule} directive.}
\usage{
renderCallModuleDirective(name, id, parameters)
}
\arguments{
\item{name}{Module name.}

\item{id}{An id string that corresponds with the id used to call the module's UI function.}

\item{parameters}{Server parameters used when calling the module.}
}
\value{
Rendered \code{callModule} directive.
}
\description{
Render a \code{callModule} directive.
}
\author{
Damian Rodziewicz
}
