% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCallModuleDirective}
\alias{renderCallModuleDirective}
\title{Render a callModule directive.}
\usage{
renderCallModuleDirective(name, id, parameters)
}
\arguments{
\item{name}{Module name.}

\item{id}{Module id.}

\item{parameters}{Server parameters used when calling the module.}
}
\value{
Rendered callModule directive.
}
\description{
Render a callModule directive.
}
\author{
Damian Rodziewicz
}
