% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderCallModuleDirectives}
\alias{renderCallModuleDirectives}
\title{Render \code{callModule} directives for provided modules.}
\usage{
renderCallModuleDirectives(modules)
}
\arguments{
\item{modules}{\code{tibble} with modules metadata.
\code{tibble} format: \code{type}, \code{name}, \code{id}, \code{parameters}.}
}
\value{
Rendered \code{callModule} directives.
}
\description{
Render \code{callModule} directives for provided modules.
}
\note{
\code{spades_simInit} and \code{spades_expt} modules are omitted,
      as they are rendered separately using \code{\link{renderSimInitDirective}}
      and \code{\link{renderExperimentDirectives}}, respectively.
}
\author{
Damian Rodziewicz
}
