% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderExperimentDirectives}
\alias{renderExperimentDirectives}
\title{Render \code{callModule} directives for experiment module if provided.}
\usage{
renderExperimentDirectives(modules)
}
\arguments{
\item{modules}{\code{tibble} with modules metadata.
\code{tibble} format: \code{type}, \code{name}, \code{id}, \code{parameters}.}
}
\value{
Rendered \code{callModule} directives.
}
\description{
Additional customizations can be provided via two files: \file{pre_experiment.R}
and \file{post_experiment.R}.
}
\author{
Alex Chubaty
}
