% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderMenuItems}
\alias{renderMenuItems}
\title{Render menu items for provided layout and available modules.}
\usage{
renderMenuItems(layout, modules)
}
\arguments{
\item{layout}{\code{tibble} with layout metadata.
\code{tibble} format: \code{tabName}, \code{menuItemName}, \code{icon},
\code{moduleName}.}

\item{modules}{\code{tibble} with modules metadata.
\code{tibble} format: \code{type}, \code{name}, \code{id}, \code{parameters}.}
}
\value{
Rendered menu items.
}
\description{
Render menu items for provided layout and available modules.
}
\author{
Damian Rodziewicz
}
