% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderSimInitDirective}
\alias{renderSimInitDirective}
\title{Render \code{callModule} directives for \code{simInit} module if provided.}
\usage{
renderSimInitDirective(modules)
}
\arguments{
\item{modules}{\code{tibble} with modules metadata.
\code{tibble} format: \code{type}, \code{name}, \code{id}, \code{parameters}.}
}
\value{
Rendered \code{callModule} directives.
}
\description{
Render \code{callModule} directives for \code{simInit} module if provided.
}
\author{
Alex Chubaty
}
