% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderTabItems}
\alias{renderTabItems}
\title{Render tab items for provided layout and available modules.}
\usage{
renderTabItems(layout, modules)
}
\arguments{
\item{layout}{Tibble with layout metadata. Tibble format: tabName, menuItemName, icon, moduleName.}

\item{modules}{Tibble with modules metadata. Tibble format: type, name, uiParameters, serverParameters.}
}
\value{
Rendered tab items.
}
\description{
Render tab items for provided layout and available modules.
}
\author{
Damian Rodziewicz
}
