% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{renderTemplate}
\alias{renderTemplate}
\title{Render a template using \pkg{whisker} package.}
\usage{
renderTemplate(templatePath, data)
}
\arguments{
\item{templatePath}{Path to the template file.}

\item{data}{Named list or environment with variables that will be
used during rendering.}
}
\value{
Rendered template.
}
\description{
Render a template using \pkg{whisker} package.
}
\author{
Damian Rodziewicz
}
