% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInfo.R
\name{simInfoUI}
\alias{simInfoUI}
\alias{simInfo}
\title{\code{SpaDES} simulation information (parent) module}
\usage{
simInfoUI(id)

simInfo(input, output, session, sim)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{sim}{A \code{simList} object}
}
\description{
\code{SpaDES} simulation information (parent) module
}
\author{
Olivia Sung, Alex Chubaty, Greyson Wang
}
