% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{sliderUI}
\alias{sliderUI}
\alias{slider}
\title{Slider Module}
\usage{
sliderUI(id, ...)

slider(input, output, session)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{...}{sliderInput arguments. See ?sliderInput for reference.}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

Reactive value with current value of the slider.
}
\description{
Function \code{sliderUI} creates a shiny module UI.

Function \code{slider} creates a slider module server function.
}
\author{
Mateusz Wyszynski
}
