% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{sliderUI}
\alias{sliderUI}
\title{Slider Module}
\usage{
sliderUI(id, label = NULL, min, max, value, interval, loop = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function}

\item{label}{Character string with which the slider will be labelled. Default is NULL - no label.}

\item{min}{The minimum value that can be selected}

\item{max}{The maximum value that can be selected}

\item{value}{The initial value of the slider}

\item{interval}{Number specifying how fast animation should progress}

\item{loop}{Should the animation be a loop. Default is FALSE - no loop.}
}
\description{
Generates a slider which can animate from minimum to maximum value.
}
\author{
Mateusz Wyszynski
}
