% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-shiny-package.R
\docType{package}
\name{SpaDES.shiny-package}
\alias{SpaDES.shiny}
\alias{SpaDES.shiny-package}
\title{The \pkg{SpaDES.shiny} package}
\description{
Create 'shiny' app interfaces for 'SpaDES' models.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://spades-shiny.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.shiny}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.shiny/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@NRCan-RNCan.gc.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
  \item Damian Rodziewicz
  \item Mateusz Wyszynski
}

Other contributors:
\itemize{
  \item Olivia Sung [contributor]
  \item Alex Tso [contributor]
  \item Greyson Wang [contributor]
  \item His Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
