% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesofRasters.R
\name{timeSeriesofRastersUI}
\alias{timeSeriesofRastersUI}
\alias{timeSeriesofRasters}
\title{Time series of rasters shiny module}
\usage{
timeSeriesofRastersUI(id)

timeSeriesofRasters(input, output, session, rctRasterList, rctUrlTemplate,
  rctPolygonList, defaultPolyName = NULL, shpStudyRegionName = NULL,
  colorPalette, maxAge, zoom = 5, mapLegend = "", mapTitle = "",
  sliderTitle = "", histTitle = "", nPolygons, nRasters,
  rasterStepSize = 10, uploadOpts = list(auth = NULL, path = NULL, user =
  NULL), rctStudyArea = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rctRasterList}{A reactive that gives a list of rasters to be displayed.}

\item{rctUrlTemplate}{The reactive url template for leaflet map tiles}

\item{rctPolygonList}{Reactive list with sets of polygons to be displayed on a leaflet map.
# TODO: decribe the format of the list!}

\item{defaultPolyName}{Name of the polygon to use as the default for mapping.}

\item{shpStudyRegionName}{Name of the study area region (from \code{rctRasterList}).}

\item{colorPalette}{Colour palette to use.}

\item{maxAge}{Maximum simulation age.}

\item{zoom}{Initial leaflet zoom.}

\item{mapLegend}{The legend text to add to the leaflet map.}

\item{mapTitle}{Title to be shown above the map.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{histTitle}{Title to be shown above the histogram.}

\item{nPolygons}{The number of available polygons.}

\item{nRasters}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}

\item{uploadOpts}{A list of options for use with file uploads:
\code{auth} logical indicating whether user is authorized to upload;
\code{path} a directory path to use for file uploads;
\code{user} the current username (used for creating user-specific paths).
The default for all options is \code{NULL}, which means do not use.}

\item{studyArea}{A \code{SpatialPolygons*} object defining the whole study area region.}

\item{subStudyArea}{Optional \code{SpatialPolygons*} object defining a smaller sub study area.}
}
\value{
Invoked for the side-effect of creating a shiny UI.

Reactive polygon selected by the user with the \code{polygonChooser} module.
         Invoked for the side-effect of creating shiny server and ui components. # TODO: reword
}
\description{
A shiny module that displays rasters changing in time on a predefined map,
with the user choosing which raster should be displayed by using the slider.
A histogram summary for each raster choice is also shown.
}
\author{
Mateusz Wyszynski

Alex Chubaty
}
