% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesofRasters.R
\name{timeSeriesofRastersUI}
\alias{timeSeriesofRastersUI}
\alias{timeSeriesofRasters}
\title{Time series of rasters shiny module}
\usage{
timeSeriesofRastersUI(id)

timeSeriesofRasters(input, output, session, rasterList, rctPolygonList,
  defaultPolyName = NULL, shpStudyRegionName = NULL, colorTable, palette,
  maxAge, zoom = 5, sim = NULL, mapLegend = "", mapTitle = "",
  sliderTitle = "", histTitle = "", nPolygons, nRasters,
  rasterStepSize = 10)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rasterList}{List of rasters to be displayed.}

\item{rctPolygonList}{Reactive list of polygons. # TODO: improve docs}

\item{defaultPolyName}{Name of the polygon to use as the default for mapping.}

\item{shpStudyRegionName}{Name of the study area region (from \code{rasterList})..}

\item{colorTable}{File that contains colour values for tiles (passed to \code{\link{gdal2Tiles}}).}

\item{palette}{Color palette for the rasters.}

\item{maxAge}{Maximum simulation age.}

\item{zoom}{Initial leaflet zoom.}

\item{sim}{A \pkg{SpaDES} simulation object (\code{simList}).}

\item{mapLegend}{The legend text to add to the leaflet map.}

\item{mapTitle}{Title to be shown above the map.}

\item{sliderTitle}{Title to be shown above the slider.}

\item{histTitle}{Title to be shown above the histogram.}

\item{nPolygons}{The number of available polygons.}

\item{nRasters}{The number of available rasters.}

\item{rasterStepSize}{Size of step in the raster slider.}
}
\value{
Invoked for the side-effect of creating a shiny UI.

Reactive polygon selected by the user with the \code{polygonChooser} module.
         Invoked for the side-effect of creating shiny server and ui components. # TODO: reword
}
\description{
A shiny module that displays rasters changing in time on a predefined map,
with the user choosing which raster should be displayed by using the slider.
A histogram summary for each raster choice is also shown.
}
\author{
Mateusz Wyszynski

Alex Chubaty
}
