% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesofRasters.R
\name{timeSeriesofRastersUI}
\alias{timeSeriesofRastersUI}
\alias{timeSeriesofRasters}
\title{Time-Since-Fire shiny module}
\usage{
timeSeriesofRastersUI(id, rastersNumber)

timeSeriesofRasters(input, output, session, rasters, polygonsList,
  shpStudyRegionFull, colorTableFile, timeSeriesofRastersPalette, maxAge,
  leafletZoomInit = 5, studyArea = "SMALL", sim = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module server function.}

\item{rastersNumber}{How many rasters can we choose from.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{rasters}{Set of rasters to be displayed.}

\item{polygonsList}{List with sets of polygons. Each such set can be displayed on a leaflet map.}

\item{shpStudyRegionFull}{Study area region.}

\item{colorTableFile}{File that contains color values for tiles.}

\item{timeSeriesofRastersPalette}{Color palette for time since fire.}

\item{maxAge}{Maximum simulation age.}

\item{leafletZoomInit}{Initial leaflet zoom.}

\item{studyArea}{Size of study area. Options: \code{"FULL"}, \code{"EXTRALARGE"},
\code{"LARGE"}, \code{"MEDIUM"}, \code{"NWT"}, \code{"SMALL"}.}

\item{sim}{A SpaDES simulation object (\code{simList}).}
}
\value{
None. Invoked for the side-effect of creating a shiny UI.

None. Invoked for the side-effect of creating a shiny server part.
}
\description{
A shiny module showing the time-since-fire values from a raster.

Function \code{timeSeriesofRasters} creates a shiny module server function
             which displays rasters changing in time on a predefined map.
             A slider, on which you can choose which raster should be currently displayed,
             is also created. Moreover, a histogram summary for each raster choice is shown.
}
\author{
Mateusz Wyszynski
}
