% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSinceFire.R
\name{timeSinceFire}
\alias{timeSinceFire}
\title{Time Since Fire Shiny Module}
\usage{
timeSinceFire(input, output, session, rasters, leafletZoomInit, studyArea)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny server session object}

\item{rasters}{Set of rasters to be displayed}

\item{leafletZoomInit}{Initial leaflet zoom}

\item{studyArea}{Size of study area. Options: "FULL", "EXTRALARGE", "LARGE", "MEDIUM", "NWT", "SMALL"}
}
\description{
This creates a shiny module which displays rasters changing in time on a predefined map. A slider is also created on which you can choose which raster should be currently displayed. Moreover, a histogram summary for each raster choice is shown.
}
\author{
Mateusz Wyszynski
}
