% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unsuspendModule}
\alias{unsuspendModule}
\title{Unsuspend hidden modules}
\usage{
unsuspendModule(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Module id.}

\item{session}{The current shiny server session.}
}
\description{
Modified from \url{https://groups.google.com/forum/#!topic/shiny-discuss/F-EwQGELS1c}.
}
\author{
Andrew Sali and Joe Cheng
}
