% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePolygons.R
\name{visualizePolygonsUI}
\alias{visualizePolygonsUI}
\alias{visualizePolygons}
\title{Visualize SpatialPolygons* Module}
\usage{
visualizePolygonsUI(id, ...)

visualizePolygons(input, output, session, polygons, proxy = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{Additional parameters passed to \code{link[leaflet]{leafletOutput}}.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{polygons}{Reactive value with polygons to visualize.}

\item{proxy}{Proxy to a leaflet map on which polygons should be displayed.
See \code{\link[leaflet]{leafletProxy}}}
}
\value{
None. Invoked for the side-effect of generating UI for a leaflet map.

None. Invoked for the side-effect of rendering leaflet map.
}
\description{
Shiny module which visualizes SpatialPolygons* objects on a leaflet map.
}
