% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeRaster.R
\name{visualizeRasterUI}
\alias{visualizeRasterUI}
\alias{visualizeRaster}
\title{Visualize raster (shiny module)}
\usage{
visualizeRasterUI(id, ...)

visualizeRaster(input, output, session, raster, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{Additional UI parameters passed to \code{link[shiny]{plotOutput}},
or server parameters passed to \code{link[rasterVis]{levelplot}}.}

\item{input}{Shiny server input object.}

\item{output}{Shiny server output object.}

\item{session}{Shiny server session object.}

\item{raster}{Raster which should be visualized.
Might be a reactive value containing raster.}
}
\value{
None. Invoked for the side-effect of rendering bar plot.
}
\description{
Display a level plot for a raster.
}
